+++ Examples/IOx/IOxExample.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using InfluxDB3.Client;
 using InfluxDB3.Client.Query;
@@ -40,6 +41,18 @@ public class IOxExample
         {
             Console.WriteLine("{0,-30}{1,-15}{2,-15}", row[0], row[1], row[2]);
         }
+
+        //
+        // Query by parametrized SQL
+        //
+        const string sqlParams = "select time,location,value from temperature where location=$location order by time desc limit 10";
+        Console.WriteLine("Query by parametrized SQL");
+        Console.WriteLine("{0,-30}{1,-15}{2,-15}", "time", "location", "value");
+        await foreach (var row in client.Query(query: sqlParams, namedParameters: new Dictionary<string, object> { { "location", "west" } }))
+        {
+            Console.WriteLine("{0,-30}{1,-15}{2,-15}", row[0], row[1], row[2]);
+        }
+
 
         //
         // Query by InfluxQL